refactor(listing): refactor listing file class
