+++ tests/ozpcenter/recommend/test_recommend.py
@@ -10,7 +10,7 @@ from rest_framework.test import APITestCase
 
 from ozpcenter import model_access as generic_model_access
 from ozpcenter.scripts import sample_data_generator as data_gen
-from tests.ozpcenter.helper import ListingFile
+from tests.ozpcenter.data_util import ListingFile
 from tests.ozpcenter.helper import APITestHelper
 from tests.ozpcenter.util import shorthand_dict
 from ozpcenter.api.listing import model_access_es
