+++ tests/ozpcenter_api/test_api_listing_search.py
@@ -9,7 +9,7 @@ from rest_framework.test import APITestCase
 
 from ozpcenter import model_access as generic_model_access
 from ozpcenter.scripts import sample_data_generator as data_gen
-from tests.ozpcenter.helper import ListingFile
+from tests.ozpcenter.data_util import ListingFile
 from tests.ozpcenter.helper import validate_listing_map_keys
 from tests.ozpcenter.helper import validate_listing_map_keys_list
 from tests.ozpcenter.helper import validate_listing_search_keys_list
